requirements.txt
flask 
flask-cors 
flask-sqlalchemy
werkzeug
PyJWT